Split app.py in smaller files
