+++ wizard/streamlit/core/__init__.py
@@ -1,0 +1,1 @@
+foo = "42"
