+++ wizard/streamlit/views/st_utils.py
@@ -1,0 +1,5 @@
+DF_HEIGHT = 300
+
+
+def needed_field(text: str) -> str:
+    return f"{text}:red[*]"
